[AnalysisQC] Align checker
* at the same time, move code related to config and utility to separate
  file
